Add a RouteLayer and RouteControl
Add the leaflet-textpath submodule as it is used in the RouteLayer
Improve the bus route extraction:
 - Now supports polygons for platforms
 - Supports routes not in a route master
 - Filter out all but one U1C route (in an attempt to make the interface less
   confusing)
Add an example using the RouteLayer and RouteControl.
Hard code the centers for some "University" buildings on the Isle of White and
in Basingstoke, such that workstations listed as in these buildings appear on
the map.